context-retrieval-state/projects/project-2.1-docuchat/requirements.txt
langchain-core>=0.3.0
langchain-chroma>=0.2.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
pydantic>=2.7.0
python-dotenv>=1.0.1
pypdf>=4.2.0
